+++ NEWS.md
@@ -4,6 +4,7 @@
 
 * Fix progress messages in `.convert_anndata_df()`
 * Allow `data.frames` in `varm` in `SCE2AnnData()`
+* Standardise `uns` names to match R conventions in `AnnData2SCE()`
 
 ## zellkonverter 1.5.3 (2022-04-19)